move content tree logic to app/helpers/content_tree.py
build_content_tree, tree splitting, slicing, and offset helpers were
pure functions with no model state — extracted to a standalone module.
Article model now imports and delegates to them.
